refactor(core,topographer): relocate topo-store public API to topographer (#369)
## Summary

Per ADR-0042, durable graph artifacts belong in `@ontrails/topographer`. The topo store is exactly that — a read-only projection of a *persisted snapshot* that writes rows to `.trails/trails.db`, tracks history across builds, and feeds lockfile exports. None of it runs to dispatch a trail. This PR moves the topo-store public API and its internal substrate out of `@ontrails/core` and into `@ontrails/topographer`.

**Breaking pre-1.0 beta change.** Every consumer that previously did `import { topoStore, createTopoStore, ... } from '@ontrails/core'` now imports from `@ontrails/topographer`. After this lands, core no longer ships any persistence-shaped API.

## What changed

**Moved out of core into topographer:**
- `packages/core/src/topo-store.ts` → `packages/topographer/src/topo-store.ts`
- `packages/core/src/internal/topo-store{,-read,-snapshots}.ts` → `packages/topographer/src/internal/...`
- Topo-store test files plus a split-out topo-snapshots suite that previously lived inside `packages/core/src/__tests__/trails-db.test.ts`.

**Stayed in core (per ADR-0014):** the generic trails-db helpers — `openReadTrailsDb`, `openWriteTrailsDb`, `ensureSubsystemSchema`, `deriveTrailsDbPath`, `deriveTrailsDir`. Topographer imports these from core. Core does NOT import topographer.

**Newly public from core:** `activationSourceKey`, `projectActivationSourceDeclaration`, `activationSourceDeclarationSignature`, and `ActivationSourceProjection`. These were single-consumer pre-existing internals; promoting them avoids reaching into core internals from topographer.

**Consumers updated:** `apps/trails/src/trails/{survey,topo-store-support,topo-support,dev-support}.ts`, `packages/warden/src/drift.ts` and its test.

**Docs:** `docs/topo-store.md`, `docs/topo-store-reference.md`, `docs/architecture.md`, `docs/api-reference.md`, `packages/topographer/README.md` (expanded), `packages/core/README.md` (migration section).

**Changeset:** `.changeset/topo-store-relocation.md` — core major, topographer minor, warden + trails patch.

Verified: typecheck (21/21), build (20/20), full repo tests green (core 1051/1051; topographer 84/84), lint (22/22), `format:check`, `check`.

## Stack

Built on #368 (the rename must land first because this relocation moves files INTO the renamed `packages/topographer/` directory). Built on by #370 and #371.

Closes TRL-610
