+++ packages/topographer/src/index.ts
@@ -26,3 +26,39 @@ export type {
   WriteOptions,
   ReadOptions,
 } from './types.js';
+
+// Topo-store public API. Persistence layer for the resolved topo graph; relies
+// on `@ontrails/core` for primitive types and the generic `trails-db` helpers.
+// See ADR-0042 for the core/topographer boundary doctrine.
+export {
+  createTopoSnapshot,
+  createMockTopoStore,
+  createTopoStore,
+  listTopoSnapshots,
+  pinTopoSnapshot,
+  topoStore,
+  unpinTopoSnapshot,
+} from './topo-store.js';
+export type {
+  CreateTopoSnapshotInput,
+  ListTopoSnapshotsOptions,
+  MockTopoStoreSeed,
+  ReadOnlyTopoStore,
+  TopoSnapshot,
+  TopoStoreExportRecord,
+  TopoStoreResourceRecord,
+  TopoStoreRef,
+  TopoStoreTrailDetailRecord,
+  TopoStoreTrailRecord,
+} from './topo-store.js';
+export {
+  countPinnedSnapshots,
+  countPrunableSnapshots,
+  countTopoSnapshots,
+  pruneUnpinnedSnapshots,
+} from './internal/topo-snapshots.js';
+export {
+  createTopoSnapshot as createStoredTopoSnapshot,
+  getStoredTopoExport,
+} from './internal/topo-store.js';
+export type { StoredTopoExport } from './internal/topo-store.js';
